+++ all-is-cubes-desktop/src/record/rmain.rs
@@ -1,5 +1,4 @@
 use std::fmt;
-use std::sync::Arc;
 use std::time::Duration;
 
 use all_is_cubes::character::{self, Character};
@@ -37,9 +36,8 @@ where
     // override it if they do want to record the UI.
     dsession
         .session
-        .graphics_options_mut()
-        .update_mut(|graphics_options| {
-            let graphics_options = Arc::make_mut(graphics_options);
+        .settings()
+        .mutate_graphics_options(|graphics_options| {
             graphics_options.show_ui = false;
             graphics_options.debug_info_text = false;
         });