+++ all-is-cubes-ui/src/apps/mod.rs
@@ -6,5 +6,8 @@ pub use input::*;
 mod session;
 pub use session::*;
 
+mod settings;
+pub use settings::*;
+
 mod time;
 pub use time::*;